pyproject.toml
[project]
name = "mathly-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "csaba szabo",email = "csabattilaszabo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "firebase-admin (>=6.8.0,<7.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "apispec (>=6.8.2,<7.0.0)",
    "marshmallow (>=4.0.0,<5.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "flask-smorest (>=0.46.1,<0.47.0)",
    "flask-cors (>=6.0.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
